aws-lambda/aws-lambda-fastapi/requirements.txt
boto3
fastapi
mangum
psycopg2-binary
sqlmodel
sqlalchemy
